// src/app/core/models/employees.ts
export class Employee {
  id?: string;
  firstName?: string;
  lastName?: string;
  state?: string;
  city?: string;
  zip?: string;
  mobileNumber?: string;
  email?: string;
  reference?: string;
  address?: string;
  dateOfBirth?: string;
  gender?: string;
  emergencyContactNumber?: string;
  relationshipWithEmp?: string;
  bankName?: string;
  accountNumber?: string;
  ifscCode?: string;
  status?: string;
  joiningDate?: string;
  lastDate?: string;
  role?: string;
  responsibility?: string;
  leaveTotal?: string;
  remainingLeaves?: string;
}